requirements.txt
slack
python-dotenv
slack_sdk
flask
flask_restx